# typescript/packages/apps/regions/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ARCADE Regions",
    "description": "Manages hierarchical structures of groups, regions, and zones.",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "def-0": {
        "description": "any non-searchable or filterable key:val attributes to specify metadata such as label, description etc.",
        "type": "object",
        "title": "attributes",
        "additionalProperties": {}
      },
      "def-1": {
        "description": "User-defined searchable tags",
        "type": "object",
        "title": "tags",
        "additionalProperties": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ]
        }
      },
      "def-2": {
        "description": "Coordinates defining a polygon.",
        "type": "array",
        "items": {
          "description": "Longitude",
          "type": "array",
          "items": {
            "description": "Latitude",
            "type": "number"
          }
        },
        "title": "polygon"
      },
...
      "def-12": {
        "type": "object",
        "properties": {
          "name": {
            "description": "The name of the Zone.",
            "type": "string"
          },
          "boundary": {
            "description": "The boundary of the Zone.",
            "$ref": "#/components/schemas/def-2"
          },
          "exclusions": {
            "description": "Boundaries to be excluded from the Zone.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/def-2"
            }
          },
          "attributes": {
            "$ref": "#/components/schemas/def-0"
          },
          "tags": {
            "$ref": "#/components/schemas/def-1"
          }
        },
        "title": "editZoneRequestBody"
      },
...
      "def-17": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Unique State ID.",
            "type": "string"
          },
          "zoneId": {
            "description": "Unique Zone ID.",
            "type": "string"
          },
          "timestamp": {
            "description": "Date/time created",
            "format": "date-time",
            "type": "string"
          },
          "attributes": {
            "$ref": "#/components/schemas/def-0"
          },
          "tags": {
            "$ref": "#/components/schemas/def-1"
          },
          "createdBy": {
            "description": "ID of owner.",
            "type": "string"
          },
          "createdAt": {
            "description": "Date/time created",
            "format": "date-time",
            "type": "string"
          },
          "updatedBy": {
            "description": "Last ID of user who made a change.",
            "type": "string"
          },
          "updatedAt": {
            "description": "Date/time updated",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "zoneId",
          "timestamp",
          "createdBy",
          "createdAt"
        ],
        "title": "stateResource"
      },
...
    }
  },
  "paths": {
...
    "/zones/{zoneId}": {
      "get": {
        "operationId": "getZone",
        "summary": "Retrieve a zone.",
        "tags": [
          "Zones"
        ],
        "description": "Retrieve the details of a zone.\n\nPermissions:\n- Only `reader` and above may view zones.\n",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "zoneId",
            "required": true,
            "description": "Unique Zone ID."
          },
          {
            "schema": {
              "default": "1.0.0",
              "type": "string"
            },
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version"
          },
          {
            "schema": {
              "default": "application/json",
              "type": "string"
            },
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type"
          }
        ],
        "x-security-scopes": [
          "admin",
          "contributor",
          "reader"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "type": "object",
                  "properties": {
                    "id": {
                      "description": "Unique Zone ID.",
                      "type": "string"
                    },
                    "name": {
                      "description": "The name of the Zone.",
                      "type": "string"
                    },
                    "regionId": {
                      "description": "Unique Region ID.",
                      "type": "string"
                    },
                    "boundary": {
                      "description": "The boundary of the Zone.",
                      "$ref": "#/components/schemas/def-2"
                    },
                    "exclusions": {
                      "description": "Boundaries to be excluded from the Zone.",
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-2"
                      }
                    },
                    "area": {
                      "description": "The area of the Zone.",
                      "type": "number"
                    },
                    "attributes": {
                      "$ref": "#/components/schemas/def-0"
                    },
                    "tags": {
                      "$ref": "#/components/schemas/def-1"
                    },
                    "state": {
                      "$ref": "#/components/schemas/def-17"
                    },
                    "createdBy": {
                      "description": "ID of owner.",
                      "type": "string"
                    },
                    "createdAt": {
                      "description": "Date/time created",
                      "format": "date-time",
                      "type": "string"
                    },
                    "updatedBy": {
                      "description": "Last ID of user who made a change.",
                      "type": "string"
                    },
                    "updatedAt": {
                      "description": "Date/time updated",
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "regionId",
                    "boundary",
                    "area",
                    "createdBy",
                    "createdAt"
                  ]
                },
                "examples": {
                  "Existing zone": {
                    "summary": "Zone retrieved successfully.",
                    "value": {
                      "id": "76ghytgt5",
                      "name": "Field 1",
                      "boundary": [
                        [
                          -104.5079674,
                          39.9194752
                        ],
                        [
                          -104.4894065,
                          39.9193435
                        ],
                        [
                          -104.4893912,
                          39.9122295
                        ],
                        [
                          -104.5078877,
                          39.9123941
                        ],
                        [
                          -104.5079674,
                          39.9194752
                        ]
                      ],
                      "regionId": "htgdjajdhja",
                      "area": 175,
                      "createdAt": "2022-08-10T23:55:20.322Z",
                      "createdBy": "someone@somewhere.com"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Not found.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteZone",
        "summary": "Deletes a zone.",
        "tags": [
          "Zones"
        ],
        "description": "Deletes a zone.\n\nPermissions:\n- Only `admin` and above may delete zones.\n",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "zoneId",
            "required": true,
            "description": "Unique Zone ID."
          },
          {
            "schema": {
              "default": "1.0.0",
              "type": "string"
            },
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version"
          },
          {
            "schema": {
              "default": "application/json",
              "type": "string"
            },
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type"
          }
        ],
        "x-security-scopes": [
          "admin"
        ],
        "responses": {
          "204": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Forbidden.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Not found.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateZone",
        "summary": "Update a zone.",
        "tags": [
          "Zones"
        ],
        "description": "Update an existing zone.\n\nPermissions:\n- Only `admins` may update zones.\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-12"
              },
              "examples": {
                "Update zone": {
                  "summary": "Update an existing zone.",
                  "value": {
                    "name": "Field 1",
                    "boundary": [
                      [
                        -104.5079674,
                        39.9194752
                      ],
                      [
                        -104.4894065,
                        39.9193435
                      ],
                      [
                        -104.4893912,
                        39.9122295
                      ],
                      [
                        -104.5078877,
                        39.9123941
                      ],
                      [
                        -104.5079674,
                        39.9194752
                      ]
                    ]
                  }
                },
                "Changing tags": {
                  "summary": "Changing the tags of a zone.",
                  "value": {
                    "tags": {
                      "tier": "GOLD"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "zoneId",
            "required": true,
            "description": "Unique Zone ID."
          },
          {
            "schema": {
              "default": "1.0.0",
              "type": "string"
            },
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version"
          },
          {
            "schema": {
              "default": "application/json",
              "type": "string"
            },
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type"
          }
        ],
        "x-security-scopes": [
          "admin",
          "contributor"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "type": "object",
                  "properties": {
                    "id": {
                      "description": "Unique Zone ID.",
                      "type": "string"
                    },
                    "name": {
                      "description": "The name of the Zone.",
                      "type": "string"
                    },
                    "regionId": {
                      "description": "Unique Region ID.",
                      "type": "string"
                    },
                    "boundary": {
                      "description": "The boundary of the Zone.",
                      "$ref": "#/components/schemas/def-2"
                    },
                    "exclusions": {
                      "description": "Boundaries to be excluded from the Zone.",
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-2"
                      }
                    },
                    "area": {
                      "description": "The area of the Zone.",
                      "type": "number"
                    },
                    "attributes": {
                      "$ref": "#/components/schemas/def-0"
                    },
                    "tags": {
                      "$ref": "#/components/schemas/def-1"
                    },
                    "state": {
                      "$ref": "#/components/schemas/def-17"
                    },
                    "createdBy": {
                      "description": "ID of owner.",
                      "type": "string"
                    },
                    "createdAt": {
                      "description": "Date/time created",
                      "format": "date-time",
                      "type": "string"
                    },
                    "updatedBy": {
                      "description": "Last ID of user who made a change.",
                      "type": "string"
                    },
                    "updatedAt": {
                      "description": "Date/time updated",
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "regionId",
                    "boundary",
                    "area",
                    "createdBy",
                    "createdAt"
                  ]
                },
                "examples": {
                  "Existing zone updated successfully": {
                    "summary": "Existing zone created successfully.",
                    "value": {
                      "id": "0980yht42",
                      "name": "Field 2",
                      "regionId": "htgdjajdhja",
                      "boundary": [
                        [
                          -104.4895628,
                          39.9390518
                        ],
                        [
                          -104.492009,
                          39.938295
                        ],
                        [
                          -104.4926527,
                          39.9376369
                        ],
                        [
                          -104.494026,
                          39.9378015
                        ],
                        [
                          -104.4971159,
                          39.9367485
                        ],
                        [
                          -104.4993046,
                          39.9345767
                        ],
                        [
                          -104.4992188,
                          39.9332933
                        ],
                        [
                          -104.4999483,
                          39.931615
                        ],
                        [
                          -104.4996908,
                          39.926909
                        ],
                        [
                          -104.4895199,
                          39.9268103
                        ],
                        [
                          -104.4895628,
                          39.9390518
                        ]
                      ],
                      "area": 655,
                      "createdAt": "2022-08-10T23:55:20.322Z",
                      "createdBy": "someone@somewhere.com",
                      "updatedAt": "2022-08-12T07:34:23.132Z",
                      "updatedBy": "someoneelse@somewhere.com"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Bad request.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                },
                "examples": {
                  "Invalid request": {
                    "summary": "Invalid request.",
                    "value": {
                      "description": "Expected `formula` to be defined but not provided."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Forbidden.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Not found.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Conflict.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "syntaxErrors": {
                      "type": "object",
                      "properties": {
                        "charPositionInLine": {
                          "type": "integer"
                        },
                        "line": {
                          "type": "integer"
                        },
                        "msg": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "charPositionInLine",
                        "line",
                        "msg"
                      ]
                    }
                  },
                  "required": [
                    "message"
                  ]
                },
                "examples": {
                  "Name in use": {
                    "summary": "The `name` is already in use.",
                    "value": {
                      "description": "Name `xyz` already exists."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:30001"
    }
  ],
  "security": [],
...
}